+++ packages/common/global-types/index.d.ts
@@ -1,5 +1,6 @@
 /// <reference lib="es2015.promise" />
 /// <reference path="./es2015/object.d.ts" />
+/// <reference path="./player-contrib-config.d.ts" />
 /// <reference path="./kaltura-player/player-config.d.ts" />
 /// <reference path="./kaltura-player/base-plugin.d.ts" />
 /// <reference path="./kaltura-player/base-middleware.d.ts" />