+++ packages/ui/src/components/banner-container/banner-container.tsx
@@ -4,6 +4,12 @@ import * as styles from './_banner-container.scss';
 
 export interface BannerContainerProps {
   onClose: () => void;
+  theme: BannerTheme;
+}
+
+interface BannerTheme {
+  backgroundColor: string;
+  blur: string;
 }
 
 export class BannerContainer extends Component<BannerContainerProps> {
@@ -28,9 +34,17 @@ export class BannerContainer extends Component<BannerContainerProps> {
     this._logger.trace(`render component`, {
       method: 'render',
     });
+
+    const {backgroundColor, blur} = this.props.theme;
+
     return (
       <div className={styles.bannerContainerRoot}>
-        <div className={styles.bannerContainer}>
+        <div
+          style={`
+                background-color:${backgroundColor}; 
+                backdrop-filter: blur(${blur});
+             `}
+          className={styles.bannerContainer}>
           <button className={styles.closeButton} onClick={props.onClose} />
           {this.props.children}
         </div>