+++ packages/ui/src/components/kitchen-sink/kitchen-sink.tsx
@@ -4,6 +4,7 @@ import * as styles from './_kitchen-sink.scss';
 
 export interface KitchenSinkProps {
   isActive: boolean;
+  kitchenSinkConfig: KalturaPlayerContribTypes.KitchenSinkConfig;
 }
 
 export class KitchenSink extends Component<KitchenSinkProps> {
@@ -40,7 +41,20 @@ export class KitchenSink extends Component<KitchenSinkProps> {
       });
     }
 
-    return <div className={this._getClass()}>{this.props.children}</div>;
+    const {backgroundColor, blur} = this.props.kitchenSinkConfig.theme;
+
+    return (
+      <div className={this._getClass()}>
+        <div
+          className={styles.backgroundLayout}
+          style={`
+                background-color:${backgroundColor}; 
+                backdrop-filter: blur(${blur});
+             `}
+        />
+        <div className={styles.children}>{this.props.children}</div>
+      </div>
+    );
   }
 
   private _getClass(): string {