+++ packages/ui/src/components/overlay-item-container/index.ts
@@ -1,0 +1,1 @@
+export {OverlayItemContainer} from './overlay-item-container';