+++ packages/ui/src/kitchen-sink-item.tsx
@@ -19,6 +19,7 @@ export interface KitchenSinkItemOptions {
   activate: (item: KitchenSinkItem) => void;
   deactivate: (item: KitchenSinkItem) => void;
   eventManager: EventsManager<KitchenSinkEvents>;
+  kitchenSinkConfig: KalturaPlayerContribTypes.KitchenSinkConfig;
 }
 
 export interface KitchenSinkItemRenderProps {
@@ -112,7 +113,7 @@ export class KitchenSinkItem {
     this._options.deactivate(this);
   }
 
-  public destroy(): void {
+  public _destroy(): void {
     if (this._isDestroyed()) {
       return;
     }
@@ -135,6 +136,7 @@ export class KitchenSinkItem {
     }
 
     const {renderContent, label} = this._options.data;
+    const {kitchenSinkConfig} = this._options;
 
     return (
       <ManagedComponent
@@ -144,6 +146,7 @@ export class KitchenSinkItem {
           <KitchenSink
             children={renderContent(props)}
             isActive={this.isActive()}
+            kitchenSinkConfig={kitchenSinkConfig}
           />
         )}
         isShown={() => !this._destroyed}