+++ packages/ui/src/kitchen-sink-manager.tsx
@@ -7,11 +7,17 @@ import {
 import {KitchenSinkItem, KitchenSinkItemRenderProps} from './kitchen-sink-item';
 import {UpperBarManager} from './upper-bar-manager';
 import {PresetManager} from './preset-manager';
-import {EventsManager, PlayerContribRegistry} from '@playkit-js-contrib/common';
+import {
+  ArrayUtils,
+  EventsManager,
+  ObjectUtils,
+  PlayerContribRegistry,
+} from '@playkit-js-contrib/common';
 import {PresetNames} from './preset-item-data';
 import {KitchenSinkContainer} from './components/kitchen-sink-container/kitchen-sink-container';
 import {KitchenSinkAdapter} from './components/kitchen-sink-adapter';
 import {ManagedComponent} from './components/managed-component';
+import {UpperBarItem} from './upper-bar-item';
 
 export interface KitchenSinkManagerOptions {
   corePlayer: KalturaPlayerTypes.Player;
@@ -40,9 +46,19 @@ const ResourceToken = 'KitchenSinkManager-v1';
 
 interface KitchenSinkPanel {
   ref: ManagedComponent | null;
-  items: KitchenSinkItem[];
+  items: {
+    kitchenSinkItem: KitchenSinkItem;
+    upperBarItem: UpperBarItem;
+  }[];
   activeItem: KitchenSinkItem | null;
 }
+
+const DefaultKitchenSinkConfig: KalturaPlayerContribTypes.KitchenSinkConfig = {
+  theme: {
+    backgroundColor: 'rgba(0, 0, 0, .8)',
+    blur: '16px',
+  },
+};
 
 export class KitchenSinkManager {
   static fromPlayer(
@@ -65,6 +81,7 @@ export class KitchenSinkManager {
 
   private _options: KitchenSinkManagerOptions;
   private _kitchenSinkAdapterRef: KitchenSinkAdapter | null = null;
+  private _kitchenSinkConfig: KalturaPlayerContribTypes.KitchenSinkConfig;
 
   on: EventsManager<KitchenSinkEvents>['on'] = this._events.on.bind(
     this._events
@@ -75,6 +92,19 @@ export class KitchenSinkManager {
 
   constructor(private options: KitchenSinkManagerOptions) {
     this._options = options;
+
+    const playerConfig =
+      options.corePlayer &&
+      options.corePlayer.config &&
+      options.corePlayer.config.contrib &&
+      options.corePlayer.config.contrib.ui &&
+      options.corePlayer.config.contrib.ui.kitchenSink
+        ? options.corePlayer.config.contrib.ui.kitchenSink
+        : {};
+    this._kitchenSinkConfig = ObjectUtils.mergeDefaults<
+      KalturaPlayerContribTypes.KitchenSinkConfig
+    >({}, DefaultKitchenSinkConfig, playerConfig);
+
     this.options.presetManager.add({
       label: 'kitchen-sink-right',
       fillContainer: true,
@@ -111,31 +141,47 @@ export class KitchenSinkManager {
       activate: this._activateItem,
       deactivate: this._deactivateItem,
       eventManager: this._events,
+      kitchenSinkConfig: this._kitchenSinkConfig,
     };
 
     const relevantPanel = this._panels[data.position];
-    const item = new KitchenSinkItem(itemOptions);
-    relevantPanel.items.push(item);
+    const kitchenSinkItem = new KitchenSinkItem(itemOptions);
+    const upperBarItem = this.options.upperBarManager.add({
+      label: data.label,
+      renderItem: data.renderIcon,
+      onClick: () => this._toggle(kitchenSinkItem),
+    });
+    relevantPanel.items.push({
+      kitchenSinkItem,
+      upperBarItem,
+    });
     if (relevantPanel.ref) {
       relevantPanel.ref.update();
     }
 
-    this.options.upperBarManager.add({
-      label: data.label,
-      renderItem: data.renderIcon,
-      onClick: () => this._toggle(item),
-    });
-
-    return item;
+    return kitchenSinkItem;
   }
 
-  //todo [sa] implement
   remove(item: KitchenSinkItem): void {
-    //deactivate item if needed
-    //remove upper bar icon
-    //remove from _items
-    //call item.destroy
-    //refresh UI ???
+    const relevantPanel = this._panels[item.data.position];
+    const itemsIndex = ArrayUtils.findIndex(
+      relevantPanel.items,
+      ({kitchenSinkItem}) => {
+        return item === kitchenSinkItem;
+      }
+    );
+
+    if (itemsIndex === -1) return;
+
+    this._deactivateItem(item);
+    const {upperBarItem} = relevantPanel.items[itemsIndex];
+    item._destroy();
+    this.options.upperBarManager.remove(upperBarItem);
+    relevantPanel.items.splice(itemsIndex, 1);
+
+    if (relevantPanel.ref) {
+      relevantPanel.ref.update();
+    }
   }
 
   private _toggle = (item: KitchenSinkItem): void => {
@@ -215,11 +261,11 @@ export class KitchenSinkManager {
   }
 
   private _renderKitchenSink(position: KitchenSinkPositions) {
-    const items = this._panels[position].items.map(item => {
+    const items = this._panels[position].items.map(({kitchenSinkItem}) => {
       const itemProps: KitchenSinkItemRenderProps = {
-        onClose: this._deactivateItem.bind(this, item),
+        onClose: this._deactivateItem.bind(this, kitchenSinkItem),
       };
-      return item.renderContentChild(itemProps);
+      return kitchenSinkItem.renderContentChild(itemProps);
     });
     return <KitchenSinkContainer>{items}</KitchenSinkContainer>;
   }