+++ scripts/bump-version.js
@@ -26,8 +26,8 @@ const args = process.argv.splice(2);
     {bold git reset --hard}
     
     To commit changes to github run:
-    {bold git commit -am "chore(release): publish ${packageVersion}}"
-    {bold git tag -a v${packageVersion}} -m "v${packageVersion}}"
+    {bold git commit -am "chore(release): publish ${packageVersion}"}
+    {bold git tag -a v${packageVersion} -m "v${packageVersion}"}
     {bold git push --follow-tags}  
   `);
 })();