+++ packages/plugin/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/kaltura/playkit-js-contrib/compare/v2.2.0...v2.3.0) (2019-11-07)
+
+
+### Features
+
+**Note:** Version bump only for package @playkit-js-contrib/plugin
+
+
+
+
+
+# [2.2.0](https://github.com/kaltura/playkit-js-contrib/compare/v2.1.0...v2.2.0) (2019-11-05)
+
+
+### Features
+
+**Note:** Version bump only for package @playkit-js-contrib/plugin
+
+
+
+
+
 # [2.1.0](https://github.com/kaltura/playkit-js-contrib/compare/v2.0.1-next.0...v2.1.0) (2019-10-31)